requirements.txt
Flask>=1.0,<=1.1.2
azure-cosmos>=1.0
